Add AvatarData::_orientation to replace _bodyYaw, _bodyPitch, _bodyRoll.
Storing rotations (or orientations) as a tripple of three angles
in degrees requires a lot of CPU intensive conversions. The rest
of the code works with quaternions, therefore it's is best to
also store the avatar orientation as a quaternion.

There is code that actually talks in terms of yaw, pitch and roll,
but this code is natively slow: storing to and from settings,
and explicit accesses to MyAvatar.bodyPitch etc variables in
scripts. In the latter case a user is probably playing with
Euler rotations because he or she is unfamiliar with quaternions
and we're not expecting a lot of such accesses. Heavy duty
avatar orientation scripts should also use MyAvatar.orientation,
which now becomes faster as a result of this patch as well
since conversions to/from Euler angles are omitted (as, imho, a
scripter would expect.)