Improved Avatar::computeRotationFromBodyToWorldUp
This function is exclusively called with argument 1.0,
but worse - in normal cases only when the avatar is
oriented world-upwards (anti gravity), which causes
it to return the identity quaternion in 99.9% of the
cases.

Nevertheless, the old code calculated a dot product,
and arccosine every call - while it is called
by Avatar::getWorldAlignedOrientation which for that
reason should have been called computeWorldAlignedOrientation.

Together with the previous patch, this makes
Avatar::getWorldAlignedOrientation a lot faster.

The old code also had a few edge problems: the angle being
larger than EPSILON is not a good guarantee that the
cross product comes out well. Also, the test
if (angle < EPSILON) assumes that the value returned
by acos is never -M_PI (but instead will be +M_PI).
The patch delays calculating the acos until absolutely
necessary (after handling the special cases). Even the
dot product is only calculated when the avatar isn't
standing up right, turning this function into a real
'get' function for the normal use case.