Remove unlogical "wobbling" during avatar roll and pitch.
Since the camera rotates parallel with the avatar head,
the orientation of the camera is the same as that of the
body (assuming the head is straight up).

This means that when the avatar rolls, then the viewer
sees instead the world roll - not the avatar (just like
in RL ;).

In first person this is also the case, and even though
in that case the camera is behind the avatar and shifted
up some fixed amount, that offset should be relative to
where the head of the avatar actually is and not relative to
where-the-head-would-be-when-we-were-standing-straight-up.
The effect of doing so namely, only causes an unlogical,
counter intuitive wobbling of the (seemingly not rotating)
avatar around the center of the screen and the camera seems
to be changing the distance to the back of the avatar too
as a result of it rotation. If you don't understand what
I mean then that is exactly my point ;).